Optimize _cleanTeamParticipation
We did much more work than necessary to update `TeamParticipation` after
removing a member from a team, scanning whole branches of the membership
tree even when we could prove that they cannot possibly contribute to
determining which participations we need to remove.  Prune the search
tree more aggressively.  In the test case at hand, this takes the time
for the `SELECT` part of this query from on the order of 6-7 seconds to
on the order of 100 milliseconds.

The tests in `lp.registry.tests.test_teammembership` already exercise
quite a number of scenarios here, and this should not change any
observed behaviour other than speed.

LP: #2036865