Allow stagingsetup to work remotely
I haven't actually been able to test this yet, so this is a little
theoretical, but it should be a no-op for the existing setup (because
`LPCONFIG=staging-db ../../utilities/pgoptions.py` returns empty).

William and I think that `walblock` will prove to be unnecessary since
replication should be able to keep up on PS5, so for now let's try
without it rather than trying to figure out how to port it.