charm: Redirect old cron-control URLs
The current URLs used for cron-control on production, qastaging, and
staging are "http://cron-control.lp.internal/production.ini",
"http://cron-control.qastaging.lp.internal/qastaging.ini", and
"http://cron-control.staging.lp.internal/staging.ini" respectively.  The
`launchpad-cron-control` charm just exposes `/cron.ini`, which is
simpler.

However, to migrate to this, we're going to need to change DNS for
`cron-control.lp.internal` etc., and it's going to be awkward to
coordinate that change with also changing all the URLs to `/cron.ini`.
Redirecting the old URLs for a transitional period will simplify this.