Remove/rename unused loop variables
`flake8-bugbear` points out:

  B007: Loop control variable not used within the loop body. If this is
  intended, start the name with an underscore.

In most cases I just changed the variable name to `_` to ignore it.  In
some cases there was already a `from lp import _` that would have been
shadowed by that approach, so I added a prefix instead.  In a few cases
the code could be simplified in other ways, or the loop variable wasn't
actually unused because `flake8-bugbear` was confused by the use of
`vars()`.