loggerhead: Set Gunicorn max_requests to 2000
We've found with Launchpad itself that the application server leaks
memory over time.  This is surely a bug somewhere, but it's proven
extremely difficult to track down, and a cheap and very effective
workaround has been to set `max_requests` in Gunicorn.  (See the "Memory
leaks" section of
https://www.chiark.greenend.org.uk/~cjwatson/blog/lp-python3.html for
some background.)

Judging by memory graphs, Loggerhead is behaving in very much the same
way, so install the same workaround there.