charm: Fix derived.archive.canonical.com vhost config
The document root for derived.archive.canonical.com points to an
"archives" subdirectory that includes only some carefully-constructed
symlinks that point into sub-sub-directories of `archives_dir`: for
example, `archives/ubuntu-rtm` is a symlink to
`ubuntu-rtm-archive/ubuntu-rtm` (relative to `archives_dir`), which lets
us expose the actual archive contents without exposing all the temporary
data and so on under other subdirectories of `ubuntu-rtm-archive`.

This is all quite historically weird, but let's mimic it so that we can
migrate from the pre-charmed deployment while preserving existing URLs.