Fix pruning of DB branch builds from Swift
It makes no sense to prune DB builds when they're older than what's on
staging, and I don't know what I was thinking when I wrote that; staging
is updated frequently, so in practice what this means is that we prune a
bunch of builds that are newer than production.  What we actually need
to do is to prune builds that are older than the DB revision deployed to
production.

Figuring this out is unfortunately challenging, but I found what seems
to be a viable approach of fetching the deployment bundle and picking it
out from there.