Revert "Build staging db inplace due to lack of disk space"
This reverts commit 23d8c45e49f6e3338c8d6c10ce1fb55237738700 in
principle, with some further changes to work properly in a charmed
deployment.  The new staging database servers have disk space for two
copies of the database with some left over, so that should be enough to
go back to the switching strategy, meaning that staging should no longer
need to be down over the weekend.