preflight: Add Juju-style user names to FRAGILE_USERS
The preflight checks got a lot less useful when we switched to Juju
deployments, because to simplify credentials management we connect as a
per-charm user and then switch to the real target role after connecting,
but `pg_stat_activity` only shows the original user name.  For safety,
fail preflight checks if any of the Juju-style user names associated
with publishers or upload processors are connected.