Rework CI build upload series check to use DistroSeriesParent
In
https://code.launchpad.net/~cjwatson/launchpad/+git/launchpad/+merge/448906,
I had to resort to doing series name comparison because I thought that
there was no proper database-level link between the build and target
series in the cases where we're using this at present.  I've since
realized that was wrong, and the two series are linked via the
`DistroSeriesParent` table, so we can do this more neatly.