ui: ensure 'demo' background is set for (qa)staging environments
After the changes to the homepage, it is no longer displaying the 'demo' background in qastaging due to a 'background-repeat: no-repeat' setting. This ensures that when we want to show this backgroud, this setting is set to 'repeat'.
The demo background makes it more obvious to users that this is a staging environment (not production)