charm: update appserver logrotate from daily to hourly
Currently log files for this service can get quite heavy (in the realm of 10 Gb) which makes it impossible to live view the latest logs of a unit. Taking into account that we run 4 appserver units, debuging an issue can lead to several minutes to download the daily log files, and the launchpad-bastion disk to get unnecessarily full.
In this change, we are rotating the logs hourly, while keeping the same amount of days worth of logs on disk. This will lead to lighter log files, and less space for logs used.