db: update snp.pro_enable schema change to add default false statement
This is relevant so that there is never a time that the value is NULL while creating the snap, which is scenario that can happen while setting the multiple snap fields when creating a new snap - if snap.pro_enable is set after snap.branch (which runs a validator that accesses the DB), then the null constraint will fail because snap.pro_enable is null (instead of false).
This change prevents any of such issues.