Allow bpr wheel without description
The previous version of the `pkginfo` library we used would set the description field to an empty string if none was found. The current version sets it to None which lead to a DB exception. This change mimics the behavior of the previous version of the `pkginfo` library by setting the description of a binary package release to an empty string if none is found

LP:#2075530