Add a "rockcraft.yaml" parser
Parse the data from *craft.yaml and use it to determine which
distroarchseries to dispatch builds to.

The current implementation follows the approach for parsing a
charmcraft.yaml file to ease and to speed up implementation, but will be
adapted to strictly parse a rockcraft.yaml with a later commit.