Add missing permissions on rockbase(arch) for rock-build-job
In 77fb7ef7c15c115ffb5e910aff0e896b38fafccb the necessary permissions
for rockbase and rockbasearch were set for most db users, but the
rock-build-job user was still missing, so the permissions could not be
added back then.

Now, that the rock-build-job user's permissions are defined, we can add
the necessary rockbase and rockbasearch permissions.