Fix API call for RockRecipe.builds
Instead of using the modern way of accessing ids, the "bare ID" was used
to preload data for rock recipes.

A year ago, in 119d2b3b00beaac8806a7340dfc70cd0b7c69d4b this was
changed, but accidentally the old code was applied here.

We do have two tests covering the preloadBuildsData method, that is
- test_getBuilds and
- test_getByBuildFarmJobs_works
but these tests did not uncover the issue.