Enable requesting a subset of architectures for rock builds
While `requestBuilds` was already exposed as an API before, it was also
necessary to expose the `architectures` param.

When this was done, a weird behaviour was noticed. Everytime a subset
of architectures was requested, an exception was raised that mentioned
that the architecture, which was defined in the rockcraft.yaml file, but
not requested, is not allowed. This makes no sense. Only a check the
other way round makes any sense.

We face a similar issue with Snap builds, see
https://bugs.launchpad.net/launchpad/+bug/2077005, but there cannot
change existing behaviour easily.

On a second glance, it made even more sense to combine the checks
whether the requested architecture is both available via the Launchpad
build farm and in the rockcraft.yaml configuration file.