Skip preloading recipe data for rock recipes
After adding support for external git repositories to rock recipes, we
experienced breakages both in the API and in the UI, as the preloading
part was not yet adapted to handle recipes with no Launchpad git
repository.

While working on a proper fix is very high priority, we need to unblock
the store team from testing rock recipes with external git repositories.

As an out of the ordinary exception, both Jürgen and Guruprasad agreed
on applying this hotfix without fixing the underlying issue for now.

The issue will be worked on the following day and must not be ignored,
as not preloading data could have severe performance impact.

This hotfix both works around issues in the API and in the UI when
preloading data, and must be reverted when the proper long term fix is
in place.

This hotfix was applied as a cowboy to qastaging and tested manually by
Jürgen Gmach.