Fix API call for CraftRecipe.builds
Instead of using the modern way of accessing ids, the "bare ID" was used
to preload data for rock recipes.

A year ago, in 119d2b3b00beaac8806a7340dfc70cd0b7c69d4b this was
changed for e.g. Charm recipes, but accidentally the old code was
applied here.

We do have two tests covering the preloadBuildsData method, that is
- test_getBuilds and
- test_getByBuildFarmJobs_works
but these tests did not uncover the issue.

We need to follow up on this and create proper test cases, but in order
to unblock the store we proceed without adding new test cases which were
not trivial to add.