upgrade lpjsmin from 0.6 to 0.7
Note that lpjsmin is only used in the make file. Thus testing involved using phony aliases within the make file in various ways especially when lpjsmin is called directly or when phony aliases using lpjsmin is called. At the end of it, the builds and tests still ran as expected thus the commit went forward.