Update txfixtures to version 0.5.2
... but only for Python 3.10 and higher, and also for a more modern
twisted version, which is not yet in place, but will be added soon.

This also means the newly required twisted version is not tested with
our current stack, as we still run Python 3.10.

So we need to start on one end, and this change is safe as we still run
3.8 in production.