Add CRLF acceptance into FileBugDataParser
FileBugDataParser is the only parser that works on ApportJobs i.e. jobs
from Apport package. Originally, only the LF line-break was expected
from Apport, and a hard-coded logic was put inside FileBugDataParser.
The problem is, this was in violation of the RFC standards for HTTP
request formats, which only accepts CRLF as its linebreaks. Because of
this, an upgrade to multipart package which parsed multipart/form-data
requests broke the apport reporting system by becoming more strict in
its parsing by refusing to parse LF line-breaks.

Thus, to upgrade multipart we need to patch apport package to only
send CRLF line-breaks, but to patch apport package we need to
make FileBugDataParser accept CRLF line-breaks in the first place.

This commit is intended to make FileBugDataParser accept CRLF while
retaining LF parsing for backwards compatibility.

To read the input blob message without changing the cursor placement in
the file-alias object I implemented a "reset" using a "close()" call
followed by an "open()" call. This was because there were no "seek()"
or "peek()" functionality in LibraryFileAlias class used in the parser

In the testing module for the FileBugDataParser, I implemented a
MockBytesIO class that inherits the built-in BytesIO class. This was
because the tests of the parser all used BytesIO as their input
stream/file, which doesn't have a open() method while having a close()
method that deletes the message from the memory. Thus, I implemented
this class so that the tests can be made with an object that acts the
same as LibraryFileAlias in our use-case.

I also added a new test text-file into the codebase to ensure we have a
message file that uses CRLF for its line-breaks so that we can test the
ApportJobs better.

Lastly, I added some unit tests for both the parser and the
ProcessApportJob class which uses the parser.