Refactor the publisher command in crontab to a bash script
crontabs have a length limit of ~1000 characters which restricts
us from only specifying a few archive references while excluding.
To workaround this restriction, instead of specifying the whole
command in the crontab itself, we refactored the command into a
helper bash script and call the bash script from the crontab
instead.