upgrade multipart with line-break parsing
- Upgraded multipart from 0.2.4 to 1.2.1
- Since apport package can only send requests with LF line-breaks, and
the newest version of "multipart" package is incapable of accepting
anything other than CRLFs; for now, we added explicit parsing of LF or
CR to CRLF until Apport package can be SRU'd to a sufficient amount of
the user-base. Added the parsing in a way that only the Apport package
and the browser's url extension of "+storeblob" can reach it.

This is a revert of a revert, reapplying commit
"a3fbbd49e00800ee6d2501ecec93ceaf3b3c5535" with a single addition:
In the added processInputs() if block, we add a more robust accessing
mechanism with ".get()" method for each condition. Without this,
webapp_publication.rst doc test fails because it doesn't put "PATH_INFO"
inside its request. Such non-existent values can also possibly happen in
production through some API calls as well, so it is best to be safe.