upgrade multipart with line-break parsing
- Upgraded multipart from 0.2.4 to 1.2.1
- Since apport package can only send requests with LF line-breaks, and
the newest version of "multipart" package is incapable of accepting
anything other than CRLFs; for now, we added explicit parsing of LF or
CR to CRLF until Apport package can be SRU'd to a sufficient amount of
the user-base. Added the parsing in a way that only the Apport package
and the browser's url extension of "+storeblob" can reach it.