Close files after they are used
The temporary files created for re-creation of HTTPInputStreams could
cause a backlog of unclosed files on the system, possibly resulting in
dangerous slow-downs. Thus we close the files that won't be used
anymore explicitly.

Also seperated the test case for the processInputs into two since
test cases with a large number of tests would result in a resource
warning during testing.

This is revert of a revert. I am reapplying the commit:
"620b8da6acdd164a53d4094488018a9e9a0f83a5"